NetworkPkg: CodeQL Fixes.
Includes changes across the repo for the following CodeQL rules:
- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest


	ShellPkg: CodeQL Fixes.

Includes changes across the repo for the following CodeQL rules:
- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest

Some false positives are filtered in CodeQlFilters.yml.

+++ NetworkPkg/UefiPxeBcDxe/PxeBcDhcp4.c
@@ -1368,7 +1368,7 @@ PxeBcDhcp4Discover (
   EFI_DHCP4_TRANSMIT_RECEIVE_TOKEN  Token;
   BOOLEAN                           IsBCast;
   EFI_STATUS                        Status;
-  UINT16                            RepIndex;
+  UINT32                            RepIndex; // MU_CHANGE - CodeQL Change - comparison-with-wider-type
   UINT16                            SrvIndex;
   UINT16                            TryIndex;
   EFI_DHCP4_LISTEN_POINT            ListenPoint;
